+++ Runtime/Subsystems/Raycast/ARRaycastFeature.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
 using UnityEngine.XR.ARSubsystems;
-
 #if UNITY_EDITOR
 using UnityEditor;
 using UnityEditor.XR.OpenXR.Features;
@@ -12,12 +11,12 @@ namespace UnityEngine.XR.OpenXR.Features.Meta
     /// Enables AR Foundation raycast support via OpenXR for Meta Quest devices.
     /// </summary>
 #if UNITY_EDITOR
-    [OpenXRFeature(UiName = "AR Foundation: Meta Quest Raycast",
+    [OpenXRFeature(UiName = "Meta Quest: AR Raycasts",
         BuildTargetGroups = new[] { BuildTargetGroup.Android },
         Company = Constants.k_CompanyName,
         Desc = "AR Foundation raycast support on Meta Quest devices",
-        DocumentationLink = Constants.k_DocumentationURL,
-        OpenxrExtensionStrings = k_OpenXRRequestedExtensions,
+        DocumentationLink = Constants.DocsUrls.k_RaycastsUrl,
+        OpenxrExtensionStrings = "",
         Category = FeatureCategory.Feature,
         FeatureId = featureId,
         Version = "0.1.0")]
@@ -28,13 +27,6 @@ namespace UnityEngine.XR.OpenXR.Features.Meta
         /// The feature id string. This is used to give the feature a well known id for reference.
         /// </summary>
         public const string featureId = "com.unity.openxr.feature.arfoundation-meta-raycast";
-
-        /// <summary>
-        /// The set of OpenXR spec extension strings to enable, separated by spaces.
-        /// For more information, refer to
-        /// <see href="https://docs.unity3d.com/Packages/com.unity.xr.openxr@1.6/manual/features.html#enabling-openxr-spec-extension-strings"/>.
-        /// </summary>
-        const string k_OpenXRRequestedExtensions = "";
 
         static List<XRRaycastSubsystemDescriptor> s_RaycastDescriptors = new();
 
@@ -56,5 +48,15 @@ namespace UnityEngine.XR.OpenXR.Features.Meta
         {
             DestroySubsystem<XRRaycastSubsystem>();
         }
+
+#if UNITY_EDITOR
+        /// <summary>
+        /// Validation Rules for ARRaycastFeature.
+        /// </summary>
+        protected override void GetValidationChecks(List<ValidationRule> rules, BuildTargetGroup targetGroup)
+        {
+            rules.AddRange(SharedValidationRules.EnableARSessionValidationRules(this));
+        }
+#endif
     }
 }